refactor ledger handling codes in main thread
